changed the process of booking a ticket, now the departure and destination stations are selected first and then the possible trains are displayed, added the linkage of schedule maodel and the route model, at the creation of the train, all the routes and schedules are created, fare calculation is now dependant on the base fare and the distance of the journey

+++ base/models.py
@@ -5,7 +5,7 @@ from users.models import User
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, null=False)
-    wallet = models.IntegerField(default=0)
+    wallet = models.DecimalField(default=0, decimal_places=2, max_digits=10)
     tickets = models.ForeignKey(Ticket, on_delete=models.DO_NOTHING, null=True)
 
     def __str__(self):
